Abstract and move filter functionality to Microsoft.DocAsCode.Metadata.ManagedReference.Common (#2468)
* Move Filters from Microsoft.DocAsCode.Metadata.ManagedReference.Roslyn to Microsoft.DocAsCode.Metadata.ManagedReference.Common.

- Does not build yet due to necessary changes that will be provided
  in following commit.

* Abstract filter functions.

- Introduce AttributeFilterData and SymbolFilterData that store information about a symbol necessary for filtering.
- Make fiter rules work with these two classes.
- Move Roslyn-specific code to RoslynFilterData.
- Move filter config file loading code into ConfigFilterRule.

* Remove extra space between method name and bracket.
