+++ src/Microsoft.DocAsCode.Metadata.ManagedReference.Common/Filters/ConfigFilterRuleExcludeItem.cs
