+++ src/Microsoft.DocAsCode.Metadata.ManagedReference.Common/Filters/ConfigFilterRuleIncludeItem.cs
