+++ src/Microsoft.DocAsCode.Metadata.ManagedReference.Common/Filters/ConfigFilterRuleItemUnion.cs
@@ -6,7 +6,6 @@ namespace Microsoft.DocAsCode.Metadata.ManagedReference
     using System;
     using System.IO;
 
-    using Microsoft.DocAsCode.DataContracts.Common;
     using YamlDotNet.Serialization;
 
     [Serializable]
@@ -49,7 +48,7 @@ namespace Microsoft.DocAsCode.Metadata.ManagedReference
             }
         }
 
-        internal ConfigFilterRuleItem Rule
+        public ConfigFilterRuleItem Rule
         {
             get
             {
